requirements.txt
streamlit~=1.41.1
langchain-openai
python-dotenv~=1.0.1
langchain-community
langchain-text-splitters
langchain-chroma
langchain-core

langchain~=0.3.14
Flask~=3.1.0